Began structuring modules importing data
Moved low-level data importing tools (timestring parsing, etc) to parsing_tools.py
Data_Importing should only be high-level (data_type switching, combining) data importing functionality
TODO: Eventually, each data type should have their own module, all located in a sub-folder.
